requirement.txt
appdirs==1.4.2
appnope==0.1.0
asyncpg==0.8.4
cffi==1.9.1
cryptography==1.7.2
decorator==4.0.11
dnspython==1.15.0
dnspython3==1.15.0
docker==2.1.0
docker-pycreds==0.2.1
Fabric3==1.13.1.post1
idna==2.2
ipython-genutils==0.1.0
packaging==16.8
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.3
pycparser==2.17
Pygments==2.2.0
pyparsing==2.1.10
python-etcd==0.4.4
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
tornado==4.4.2
traitlets==4.3.2
urllib3==1.20
wcwidth==0.1.7
websocket-client==0.40.0
